refactor the jsonrpc module into separate files
Also added a few convenience From impls, and changed errors
