+++ src/jsonrpc/test/mod.rs
@@ -1,0 +1,4 @@
+use {JsonRpc, RpcError};
+use serde_json::{Value, to_value, to_string};
+
+// TODO: Add tests
